Add DTO\S

+++ DeskManager/Services/LocationService.cs
@@ -1,6 +1,8 @@
 ﻿using DeskManager.Entities;
 using DeskManager.Exceptions;
 using DeskManager.Models;
+using DeskManager.Models.Mappers;
+using DeskManager.Models.Validators;
 using DeskManager.Services.Interfaces;
 using FluentValidation;
 using Microsoft.EntityFrameworkCore;
@@ -17,40 +19,40 @@ public class LocationService : ILocationService
         _dbContext = dbContext;
         _createLocationValidation = createLocationValidation;
     }
-    public async Task<List<Location>> GetAll()
+    public async Task<List<LocationDto>> GetAll()
     {
         var locations = await _dbContext.Locations
             .Include(l => l.Desks)
             .ToListAsync();
-        return locations;
+        return locations.Select(LocationMapper.LocationToLocationDto).ToList();
     }
 
-    public async Task<Location> GetLocation(int id)
+    public async Task<LocationDto> GetLocation(int id)
     {
         var location = await GetById(id);
-        return location;
+        return LocationMapper.LocationToLocationDto(location);
     }
 
-    public async Task<int> CreateLocation( CreateLocationDto locationName)
+    public async Task<LocationDto> CreateLocation( CreateLocationDto createLocationDto)
     {
-        var validationResult = await _createLocationValidation.ValidateAsync(locationName);
+        var validationResult = await _createLocationValidation.ValidateAsync(createLocationDto);
         if (!validationResult.IsValid)
         {
             var errors = string.Join(Environment.NewLine, validationResult.Errors);
             throw new WrongDataException($"Create location failed: {errors}");
         }
-        var location = new Location()
-        {
-            Name = locationName.Name
-        };
+
+        var location = LocationMapper.CreateLocationDtoToLocation(createLocationDto);
         await _dbContext.Locations.AddAsync(location);
         await _dbContext.SaveChangesAsync();
-        return location.Id;
+        return LocationMapper.LocationToLocationDto(location);
     }
 
     public async Task DeleteLocation(int id)
     {
         var location = await GetById( id);
+        if (location.Desks.Any())
+            throw new InvalidOperationException("this location is not empty");
         _dbContext.Locations.Remove(location);
         await _dbContext.SaveChangesAsync();
